requirements.txt
blinker==1.9.0
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
CryptPandas==1.0.1
Flask==3.1.0
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.1.3
packaging==24.2
pandas==2.2.3
plotly==5.24.1
pyarrow==18.0.0
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
scipy==1.14.1
six==1.16.0
slack_bolt==1.21.2
slack_sdk==3.33.3
tenacity==9.0.0
tzdata==2024.2
Werkzeug==3.1.3
